+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - increase xp from tasks - improves new player experience
 - increase returning player booster to 1 week
+- increase max guild level to 1000 and increase (high) levelling difficulty
 
 ## other